toggle created for log in and log out buttons
